refactor: userService의 delete함수: 조회와 삭제 하나로 묶기

+++ src/main/java/org/team14/webty/user/service/UserService.java
@@ -70,6 +70,11 @@ public class UserService {
 
 	@Transactional
 	public void delete(WebtyUser webtyUser) {
+
+		Optional<WebtyUser> opWebtyUser = userRepository.findById(webtyUser.getUserId());
+		if (opWebtyUser.isEmpty()) {
+			throw new IllegalArgumentException("존재하지 않는 유저");
+		}
 		userRepository.delete(webtyUser);
 	}
 }
